// src/app/models/customer/customer-type-three.ts
export class CustomerTypeThree {

    customerId: number;

    firstName: string;

    secondName: string;

    lastName: string;

    secondLastName: string;

    documentTypeIdMaster: number;

    document: string;

    address: string;

    departmentIdDane: string;

    cityIdDane: string;

    email: string;

    phoneNumber: string;
}